passions/requirement.txt
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.7
django-facebook==6.0.3
django-rest-framework==0.1.0
djangorestframework==3.10.3
facebook-sdk==3.1.0
html5lib==1.0.1
idna==2.8
Pillow==6.2.1
psycopg2-binary==2.8.4
pycparser==2.19
Pyphen==0.9.5
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tinycss2==1.0.2
Unidecode==1.1.1
urllib3==1.25.7
WeasyPrint==50
webencodings==0.5.1
